Validation changes

+++ app/services/role/role_service.py
@@ -3,7 +3,7 @@ from mongoengine.queryset import NotUniqueError
 from ...protos import RoleServicer, RoleMultipleResponse, RoleResponse, RoleTableResponse, RoleEmpty, add_RoleServicer_to_server
 from ...utils import parser_all_object, parser_one_object, not_exist_code, exist_code, paginate, parser_context
 from ...utils.validate_session import is_auth
-from ...models import Roles
+from ...models import Roles, Auth
 from ..bootstrap import grpc_server
 
 class RoleService(RoleServicer):
@@ -78,6 +78,10 @@ class RoleService(RoleServicer):
         try:
             auth_token = parser_context(context, 'auth_token')
             is_auth(auth_token, '00_role_delete')
+
+            user = Auth.objects(role=request.id)
+
+            if len(user) > 0: raise Exception('role_assigned')
 
             role = Roles.objects.get(id=request.id)
             role = role.delete()
